+++ gdsfactory/samples/21_add_fiber_array.py
@@ -11,7 +11,7 @@ def big_device_with_gratings() -> gf.Component:
     component = big_device(nports=10)
     radius = 5.0
     return gf.routing.add_fiber_array(
-        component=component, radius=radius, fanout_length=50.0
+        component=component, radius=radius, fanout_length=50.0, radius_loopback=10
     )