+++ gdsfactory/samples/sample_reticle.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import gdsfactory as gf
 
 
-@gf.cell(check_ports=False)
+@gf.cell
 def spiral_gc(**kwargs):
     c = gf.c.spiral(**kwargs)
     c = gf.routing.add_fiber_array(c)
@@ -13,7 +13,7 @@ def spiral_gc(**kwargs):
     return c
 
 
-@gf.cell(check_ports=False)
+@gf.cell
 def mzi_gc(length_x=10, **kwargs):
     c = gf.components.mzi2x2_2x2_phase_shifter(
         length_x=length_x, auto_rename_ports=False, **kwargs
@@ -26,7 +26,7 @@ def mzi_gc(length_x=10, **kwargs):
     return c
 
 
-@gf.cell(check_ports=False)
+@gf.cell
 def sample_reticle(grid: bool = False) -> gf.Component:
     """Returns MZI with TE grating couplers."""
     from gdsfactory.generic_tech.cells import (
@@ -34,9 +34,7 @@ def sample_reticle(grid: bool = False) -> gf.Component:
     )
 
     mzis = [mzi_gc(length_x=lengths) for lengths in [100, 200, 300]]
-
     spirals = [spiral_gc(length=length) for length in [0, 100, 200]]
-
     rings = []
     for length_x in [10, 20, 30]:
         ring = gf.components.ring_single_heater(length_x=length_x)
@@ -47,9 +45,8 @@ def sample_reticle(grid: bool = False) -> gf.Component:
         ring_te.name = f"ring_{length_x}"
         rings.append(ring_te)
 
-    copies = 3
+    copies = 3  # number of copies of each component
     components = mzis * copies + rings * copies + spirals * copies
-
     if grid:
         return gf.grid(components)
     c = gf.pack(components)